Move tests to OpenApiSchemaReferenceTransformerTests
